Don't include inactive messages in msgset generation
Currently only deleted messages that will be expunged are marked
inactive, so it shouldn't be an issue.  However, before the previous
commit (changing sorting by UID), I think there was a possibility for
a bug if trailing expunge-marked messages with MAXINT index were also
marked inactive.

The sort change fixes that, but to make the code clearer just remove
the possibility of inactive messages being included.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/b0f54225184adc35ace668393cd29e9a6f6d250c

Co-authored-by: Kevin McCarthy <kevin@8t8.us>